+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Added method to check form submitted `validator->submitted()`.
+- Added method to check validaton and form submitted `validator->passed()`.
+- Added method to check validaton and form submitted `validator->fails()`.
 
 ## [0.5.0] - 2022-02-15
 ### Added